refactor fully configured
